+++ src/main.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 )
 
-const version = "0.0.6"
+const version = "0.1.0"
 
 func main() {
 	if len(os.Args) > 1 && (os.Args[1] == "-h" || os.Args[1] == "--help") {